CA-118 split device
